cli: Move git config functions to config package
This allows them to be used by the installer
